fix: Refactoring `--tf-path` out to shared flag
